+++ lib/config/options.dart
@@ -4,14 +4,27 @@ const kInputAbbr = 'i';
 const kOutput = 'output';
 const kOutputAbbr = 'o';
 
+/// Options class to store a single option
 class Option<T> {
+  /// The name of the option
   final String name;
+
+  /// The abbreviation of the option
   final String? abbr;
+
+  /// The help text for the option when using the help flag
   final String help;
+
+  /// The default value of the option
   final T defaultValue;
 
+  /// The parsed value of the option
   T? _value;
+
+  /// Returns the value of the option, if not set, the default value is used
   T get value => _value ?? defaultValue;
+
+  /// Sets the value of the option
   set value(T value) {
     // * All string options are paths at the moment, sanitize the path
     if (value is String && value.contains('/')) {
@@ -21,10 +34,15 @@ class Option<T> {
     }
   }
 
+  /// Constructor for the option class
   Option(this.name, this.defaultValue, this.help, [this.abbr]);
 
+  /// Copies this option into a new object
   Option<T> copy() => Option<T>(name, defaultValue, help, abbr);
 
+  /// When merging options we can use the + operator to merge the values
+  /// of the options. If there are more than one option the filled in and
+  /// not the same as the default, the second option will be used.
   Option<T> operator +(Option<T> other) {
     if (name != other.name) throw 'Cannot add to options of different type';
     final newOption = copy();
@@ -38,7 +56,9 @@ class Option<T> {
   }
 }
 
+/// Container that holds all available options for this program
 class Options {
+  /// The list of all available options
   List<Option<String>> options = [
     Option<String>(
       kInput,
@@ -54,12 +74,14 @@ class Options {
     ),
   ];
 
+  /// Returns the option with the given name
   Option<T> getOption<T>(String name) {
     final results = options.where((element) => element.name == name);
     if (results.isEmpty) throw 'Cannot find option with name `$name`';
     return results.first as Option<T>;
   }
 
+  /// Sets the option with the given name to the given value
   void setOption<T>(String name, T? value) {
     if (value != null) {
       getOption<T>(name).value = value;
@@ -74,11 +96,15 @@ Options:
 ''';
   }
 
+  /// When merging options we can use the + operator to merge the values
   Options operator +(Options other) {
     final result = Options();
 
     void mergeOption(Option<dynamic> option) {
-      result.setOption(option.name, (option + other.getOption(option.name)).value);
+      result.setOption(
+        option.name,
+        (option + other.getOption(option.name)).value,
+      );
     }
 
     options.forEach(mergeOption);