+++ scripts/publish.sh
@@ -1,0 +1,5 @@
+
+dart format . --fix
+dart analyze .
+
+dart pub publish